reworking pam
